Redesign user registration flow
With the updated flow, we now perform a manual check to verify if the username already exists in the database before attempting to insert a new entity. If the username is taken, a UsernameUnavailableException will be thrown.

Additionally, any exceptions that occur while saving the entity to the database will be caught and rethrown as a UserRegistrationException.

The endpoint no longer returns user data. Instead, upon successful registration, it simply responds with a 201 Created status.