Creating user authentication flow
This introduces a new POST endpoint for user login.

The service flow begins by verifying the existence of the username; if it does not exist, a BadCredentialsException is thrown.

Next, the plain-text password is compared to the hashed password associated with the found user. If the passwords do not match, a BadCredentialsException is again thrown.

If both checks pass, the user data is returned with a 200 OK status.

In the future, the client will receive a Bearer token along with the user data, which can be used in future requests.